// src/types/types.ts
// Generated by https://quicktype.io

export interface ApiError {
  statusCode: number;
  message: string;
  error: string;
}

// Generated by https://quicktype.io

export interface Response {
  data:       Data;
  status:     number;
  statusText: string;
  headers:    ResponseHeaders;
  config:     Config;
  request:    Request;
}

interface Config {
  transitional:      Transitional;
  adapter:           string[];
  transformRequest:  null[];
  transformResponse: null[];
  timeout:           number;
  xsrfCookieName:    string;
  xsrfHeaderName:    string;
  maxContentLength:  number;
  maxBodyLength:     number;
  env:               Request;
  headers:           ConfigHeaders;
  method:            string;
  url:               string;
  data:              string;
}

interface Request {
}

interface ConfigHeaders {
  Accept:         string;
  "Content-Type": string;
}

interface Transitional {
  silentJSONParsing:   boolean;
  forcedJSONParsing:   boolean;
  clarifyTimeoutError: boolean;
}

interface Data {
  access_token: string;
}

interface ResponseHeaders {
  "content-length": string;
  "content-type":   string;
}